Fixed signedness compilation warning Fixed compilation warning due a
2006-04-14  German Poo-Caaman~o  <gpoo@ubiobio.cl>

	* src/nettool.c (netinfo_io_text_buffer_dialog):
	* src/ping.[ch] (ping_foreach):
	* src/netstat.[ch] (netstat_foreach):
	* src/lookup.[ch] (lookup_foreach):
	* src/finger.[ch] (finger_foreach):
	* src/traceroute.[ch] (traceroute_foreach): Fixed signedness compilation
	  warning
	* src/info.c (info_load_iface): Fixed compilation warning due a
	  variable unitialized

+++ src/finger.c
@@ -134,11 +134,11 @@ finger_do (Netinfo * netinfo)
 
 /* Process each line from finger command */
 void
-finger_foreach (Netinfo * netinfo, gchar * line, gint len,
+finger_foreach (Netinfo * netinfo, gchar * line, gssize len,
 		gpointer user_data)
 {
 	gchar *text_utf8;
-	gssize bytes_written;
+	gsize bytes_written;
 	GtkTextBuffer *buffer = NULL;
 	GtkTextIter iter;
