+++ src/traceroute.c
@@ -109,11 +109,11 @@ traceroute_do (Netinfo * netinfo)
 
 /* Process each line from ping command */
 void
-traceroute_foreach (Netinfo * netinfo, gchar * line, gint len,
+traceroute_foreach (Netinfo * netinfo, gchar * line, gssize len,
 		    gpointer user_data)
 {
 	gchar *text_utf8;
-	gssize bytes_written;
+	gsize bytes_written;
 	GtkTextBuffer *buffer = NULL;
 	GtkTextIter iter;
